refactor: move cache to old_cache
