+++ crates/rspack_core/src/compiler/hmr.rs
@@ -61,9 +61,9 @@ where
       let mut all_files = modified_files.clone();
       all_files.extend(removed_files.clone());
 
-      self.cache.end_idle();
+      self.old_cache.end_idle();
       self
-        .cache
+        .old_cache
         .set_modified_files(all_files.into_iter().collect());
       self.plugin_driver.resolver_factory.clear_cache();
 
@@ -73,7 +73,7 @@ where
         self.resolver_factory.clone(),
         self.loader_resolver_factory.clone(),
         Some(records),
-        self.cache.clone(),
+        self.old_cache.clone(),
         Some(ModuleExecutor::default()),
         modified_files,
         removed_files,
@@ -111,7 +111,7 @@ where
       fast_set(&mut self.compilation, new_compilation);
       self.compile().await?;
 
-      self.cache.begin_idle();
+      self.old_cache.begin_idle();
     }
 
     self.compile_done().await?;
