+++ crates/rspack_core/src/compiler/make/repair/add.rs
@@ -88,7 +88,6 @@ impl Task<MakeTaskContext> for AddTask {
       resolver_factory: context.resolver_factory.clone(),
       compiler_options: context.compiler_options.clone(),
       plugin_driver: context.plugin_driver.clone(),
-      cache: context.cache.clone(),
     })])
   }
 }
