+++ crates/rspack_core/src/compiler/make/repair/process_dependencies.rs
@@ -117,7 +117,6 @@ impl Task<MakeTaskContext> for ProcessDependenciesTask {
         loader_resolver_factory: context.loader_resolver_factory.clone(),
         options: context.compiler_options.clone(),
         plugin_driver: context.plugin_driver.clone(),
-        cache: context.cache.clone(),
         current_profile,
       }));
     }
