+++ crates/rspack_core/src/lib.rs
@@ -24,8 +24,8 @@ pub mod external_module;
 pub use external_module::*;
 mod logger;
 pub use logger::*;
-pub mod cache;
 mod normal_module;
+pub mod old_cache;
 mod raw_module;
 pub use raw_module::*;
 mod exports_info;
