+++ crates/rspack_core/src/old_cache/storage/memory.rs
